// src/features/mapaSeparacao/types/product-types.ts
export interface ProductsItem {
  skuCode: string;
  skuDescription: string;
  shelf: number;
  weightType: number;
  boxWeight: number;
  unitsPerBox: number;
  boxesPerPallet: number;
  line: string;
  redRange: number;
  orangeRange: number;
  yellowRange: number;
  greenRange: number;
  pickWay: number;
  address: string;
  empresa: string;
}