// src/features/mapaSeparacao/types/routing-type.ts
export interface RoutingItem {
  company: string;
  route: string;
  transport: string;
  licensePlate: string;
  shipmentRoute: string;
  sequence: number;
  customer: string;
  customerName: string;
  location: string;
  routedVehicle: string;
  effectiveVehicle: string;
  carrier: string;
  cargoType: string;
} 